✨ some changes of clone web
